[FIX] 게이트 웨이 필터링 문제 해결 및 CORS 설정

+++ product-service/src/main/java/org/hanghae99/productservice/controller/ProductController.java
@@ -1,10 +1,10 @@
 package org.hanghae99.productservice.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.hanghae99.productservice.dto.*;
-import org.hanghae99.productservice.dto.Order;
 import org.hanghae99.productservice.entity.Product;
 import org.hanghae99.productservice.service.ProductService;
 import org.springframework.http.HttpStatus;
@@ -22,7 +22,8 @@ public class ProductController {
     @Transactional
     @Operation(summary = "상품 생성", description = "관리자 제한")
     @PostMapping("/create")
-    public ResponseEntity<ProductResponseDto> createProduct(@RequestBody ProductRequestDto productRequestDto) {
+    public ResponseEntity<ProductResponseDto> createProduct(@RequestBody ProductRequestDto productRequestDto, HttpServletRequest request) {
+        if (!request.getHeader("x-claim-auth").equals("admin")) throw new IllegalArgumentException("관리자 권한이 아닙니다.");
         return ResponseEntity.status(HttpStatus.CREATED).body(productService.createProduct(productRequestDto));
     }
 
@@ -32,7 +33,7 @@ public class ProductController {
         return ResponseEntity.status(HttpStatus.OK).body(productService.getProducts());
     }
 
-    @Operation(summary = "상품 상세 조회")
+    @Operation(summary = "상품 상세 조회", description = "로그인 없이도 이용할 수 있습니다.")
     @GetMapping("/search/{productNo}")
     public ResponseEntity<ProductResponseDto> getProduct(@PathVariable Long productNo) {
         return ResponseEntity.status(HttpStatus.OK).body(productService.getProduct(productNo));
@@ -46,21 +47,24 @@ public class ProductController {
 
     @Operation(summary = "상품 재입고")
     @PutMapping("/{productNo}/re-stock")
-    public ResponseEntity<ProductResponseDto> reStockProduct(@PathVariable Long productNo, @RequestBody ReStockRequestDto reStockRequestDto) {
+    public ResponseEntity<ProductResponseDto> reStockProduct(@PathVariable Long productNo, @RequestBody ReStockRequestDto reStockRequestDto, HttpServletRequest request) {
+        if (!request.getHeader("x-claim-auth").equals("admin")) throw new IllegalArgumentException("관리자 권한이 아닙니다.");
         return ResponseEntity.status(HttpStatus.OK).body(productService.reStockProduct(productNo, reStockRequestDto));
     }
 
     @Transactional
     @Operation(summary = "상품 정보 수정", description = "관리자 제한")
     @PutMapping("/{productNo}")
-    public ResponseEntity<ProductResponseDto> updateProduct(@PathVariable Long productNo, @RequestBody ProductRequestDto productRequestDto) {
+    public ResponseEntity<ProductResponseDto> updateProduct(@PathVariable Long productNo, @RequestBody ProductRequestDto productRequestDto, HttpServletRequest request) {
+        if (!request.getHeader("x-claim-auth").equals("admin")) throw new IllegalArgumentException("관리자 권한이 아닙니다.");
         return ResponseEntity.status(HttpStatus.OK).body(productService.updateProduct(productNo, productRequestDto));
     }
 
     @Transactional
     @Operation(summary = "상품 삭제", description = "관리자 제한")
     @DeleteMapping("/{productNo}")
-    public ResponseEntity<ApiResponseDto> deleteProduct(@PathVariable Long productNo) {
+    public ResponseEntity<ApiResponseDto> deleteProduct(@PathVariable Long productNo, HttpServletRequest request) {
+        if (!request.getHeader("x-claim-auth").equals("admin")) throw new IllegalArgumentException("관리자 권한이 아닙니다.");
         return ResponseEntity.status(HttpStatus.OK).body(productService.deleteProduct(productNo));
     }
